feat: add detailed report workbook
