+++ .gitignore
@@ -1,2 +1,3 @@
 source/new_project/data/
 source/new_project/__pycache__/
+source/new_project/node_modules/
